Fixed errors with MBTradingService so it gives a price for market orders and also holds the initial synchronization until accounts get loaded.

+++ Platform/TickZoomAPI1.0/Classes/Elapsed.cs
@@ -76,8 +76,8 @@ namespace TickZoom.Api
 			get { return (int) ( (elapsed * MillisecondsPerDay) % MillisecondsPerSecond); }
 		}
 		
-		public int TotalSeconds {
-			get { return (int) (elapsed * SecondsPerDay); }
+		public long TotalSeconds {
+			get { return (long) (elapsed * SecondsPerDay); }
 		}
 		
 		public long TotalMilliseconds {
